fix: resolve PR concerns
